+++ src/java/Servlets/CadastrarCargo.java
@@ -44,7 +44,7 @@ public class CadastrarCargo extends HttpServlet {
         cargo.setSalario(Float.valueOf(request.getParameter("Salario")));
         cargo.setRequisitos(request.getParameter("Requisitos"));
         cargo.setCargaMinima(Integer.valueOf(request.getParameter("CargaMinima")));
-        cargo.setDescontoImpostos(Float.valueOf(request.getParameter("DescontoImpostos")));
+        cargo.setDescontoImpostos(Integer.valueOf(request.getParameter("DescontoImpostos")));
         cargoDAO.cadastrarCargo(cargo);
         RequestDispatcher rd = getServletContext().getRequestDispatcher("/manter_cargos.jsp");
         rd.forward(request, response);
